fixed warning

+++ c_extension/src/shaderIO.c
@@ -7,7 +7,7 @@ void sendToShader(struct SceneData scene, struct argData args) {
          scene.cameraPos[1], scene.cameraPos[2], scene.cameraDir[0],
          scene.cameraDir[1], scene.cameraDir[2]);
   printf("program - %d\n", scene.shaderProgram);
-  for (int i = 0; i < args.numMeshes; i++) {
+  for (uint64_t i = 0; i < args.numMeshes; i++) {
     printf("%ld %ld %ld %ld\n", scene.meshes[i].startingVertex,
            scene.meshes[i].numTriangles, scene.meshes[i].materialID,
            scene.meshes[i].objectID);
